Move subgraph types
