+++ src/services/balancer/subgraph/entities/poolSwaps/index.ts
@@ -2,9 +2,10 @@ import { getAddress } from '@ethersproject/address';
 
 import { toJsTimestamp } from '@/composables/useTime';
 import { web3Service } from '@/services/web3/web3.service';
+import { QueryBuilder } from '@/types/subgraph';
 
 import Service from '../../balancer-subgraph.service';
-import { PoolSwap, QueryBuilder } from '../../types';
+import { PoolSwap } from '../../types';
 import queryBuilder from './query';
 
 export default class PoolSwaps {
