+++ src/services/pool/types.ts
@@ -1,6 +1,11 @@
 import { BigNumber } from 'ethers';
 
-import { Address } from '@/types';
+export type QueryArgs = Record<string, any>;
+export type QueryAttrs = Record<string, any>;
+export type QueryBuilder = (
+  args?: QueryArgs,
+  attrs?: QueryAttrs
+) => Record<string, any>;
 
 export enum PoolType {
   Weighted = 'Weighted',
